Refactor .Internals into .Common, .Windows, and .Posix (tested only on Windows)
