+++ process.cabal
@@ -47,6 +47,11 @@ library
     if impl(ghc)
         exposed-modules:
             System.Process.Internals
+        other-modules: System.Process.Common
+        if os(windows)
+            other-modules: System.Process.Windows
+        else
+            other-modules: System.Process.Posix
 
     c-sources:
         cbits/runProcess.c
@@ -66,6 +71,7 @@ library
     if os(windows)
         build-depends: Win32 >=2.2 && < 2.4
         extra-libraries: kernel32
+        cpp-options: -DWINDOWS
     else
         build-depends: unix >= 2.5 && < 2.8
